made ModelMultipleChoiceField use a verbose_label in label_from_instance when present
sometimes we have found it is useful to display additional information when a model is being used as a many to many relationship rather than just its regular string representation. Using this field, the label will first be taken from 'verbose_label' method on the object when it is present, or just do the regular behavior when not.